Added the initial constructor of the 'Event' class and introduced a few visual corrections in the code

+++ src/main/java/com/infoshareacademy/Category.java
@@ -1,11 +1,11 @@
 package com.infoshareacademy;
 
 public class Category {
-    private Integer categoryID;
-    private String  categoryName;
+    private Long categoryID;
+    private String categoryName;
     private CategoryType categoryType;
 
-    public Category(Integer categoryID, String categoryName, CategoryType categoryType) {
+    public Category(Long categoryID, String categoryName, CategoryType categoryType) {
         this.categoryID = categoryID;
         this.categoryName = categoryName;
         this.categoryType = categoryType;
