+++ src/main/java/com/infoshareacademy/CategoryType.java
@@ -1,10 +1,10 @@
 package com.infoshareacademy;
 
 public class CategoryType {
-    private Integer categoryTypeID;
+    private Long categoryTypeID;
     private String categoryTypeName;
 
-    public CategoryType(Integer categoryTypeID, String categoryTypeName) {
+    public CategoryType(Long categoryTypeID, String categoryTypeName) {
         this.categoryTypeID = categoryTypeID;
         this.categoryTypeName = categoryTypeName;
     }
