+++ src/main/java/com/infoshareacademy/Place.java
@@ -1,16 +1,16 @@
 package com.infoshareacademy;
 
 public class Place {
-    private Integer placeID;
+    private Long placeID;
     private Adress adress;
     private String namePlace;
     private String subname;
 
-    public Place(Integer placeID, Adress adress, String namePlace) {
-        this(placeID, adress,namePlace,null);
+    public Place(Long placeID, Adress adress, String namePlace) {
+        this(placeID, adress, namePlace, null);
     }
 
-    public Place(Integer placeID, Adress adress, String namePlace, String subname) {
+    public Place(Long placeID, Adress adress, String namePlace, String subname) {
         this.placeID = placeID;
         this.adress = adress;
         this.namePlace = namePlace;
